Add link to exhibit and exhibit page(s) in item page
Note that some of this functionality is in the plugin/CulCustomize,
which is in its own repo. The code we added here in this commit
makes use the following helper function from CulCustomize:
cul_display_links_to_exhibit_pages_containing_item()

We use $_SERVER['HTTP_REFERER'] to get the previous page.
We test using strstr() to make sure the link
to the previous page (which should be the exhibit
page containing the item, assuming we got to the item page
via the exhibit page) contains the uri of the exhibit.