Remove unused code that checks which contact info to use
The legacy theme is used for only one exhibition:
"The Varsity Show". Therefore, the contact information will
always be the University Archives contact information, and it makes
no sense to have other options based on the exhibition title. This
theme was probably based on a previous theme, and the unused code is
probably leftover code from that original theme.
The unused/useless code was removed.